+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.8.0] - 2023-12-07
+
+### [Changelog][0.8.0-changes]
+
+### Added
+
+- Jump forward/backward by number of lines, or by time interval in microseconds, milliseconds, seconds, minutes, hours or days.
+- Support for `--inline` command-line option, to merge logs into a single column instead of side by side. (Issue #32)
+
+### Fixed
+
+- Some Python version incompatibilities in type annotations (Issue #33)
+
+
 ## [0.7.0] - 2023-10-07
 
 ### [Changelog][0.7.0-changes]
@@ -106,6 +120,7 @@
 - Interactive browser will use the screen width by default, or accept a command-line argument.
 
 
+[0.8.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.8.0
 [0.7.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.7.0
 [0.6.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.6.0
 [0.5.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.5.0
@@ -115,6 +130,7 @@
 [0.2.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.2.0
 [0.1.0]: https://github.com/ptmcg/log_merger/releases/tag/v0.1.0
 
+[0.8.0-changes]: https://github.com/ptmcg/log_merger/compare/v0.7.0...v0.8.0
 [0.7.0-changes]: https://github.com/ptmcg/log_merger/compare/v0.6.0...v0.7.0
 [0.6.0-changes]: https://github.com/ptmcg/log_merger/compare/v0.5.0...v0.6.0
 [0.5.0-changes]: https://github.com/ptmcg/log_merger/compare/v0.4.0...v0.5.0